Edit hashcode, cleanup, and move onclick flytoandfilter item logic
